refactor: recibo honorario - se agrego la vista previa del honorario y se modifico el html del formulario de datos para generar el recibo del honorario

+++ src/main/java/com/cfm/sacc/operaciones/model/ReciboHonorario.java
@@ -13,7 +13,7 @@ public class ReciboHonorario {
 	private Integer id;
 	private String clienteRFC;
 	private String nombreCliente;
-	private Double importeHonorario;
+	private Float importeHonorario;
 	private String importeLetra;
 	private String concepto;
 }
